+++ pages/success/index.tsx
@@ -7,7 +7,38 @@ import Flex from "components/Flex";
 import Section from "components/Section";
 import NextImage from "next/image";
 import pam from "./assets/pam.png";
-import LI, { TopLinks } from "./constants";
+import Icon from "components/Icon";
+import type { IconName } from "components/Icon";
+import Link from "components/Link";
+
+interface TopLink {
+  iconName: IconName;
+  href: string;
+  hrefName: string;
+}
+
+const TopLinks: Array<TopLink> = [
+  {
+    iconName: "quickstart",
+    href: "https://goteleport.com/docs/getting-started/",
+    hrefName: "Quickstart Guide",
+  },
+  {
+    iconName: "documents",
+    href: "https://goteleport.com/docs/",
+    hrefName: "Developer Docs",
+  },
+  {
+    iconName: "shieldCheck",
+    href: "https://goteleport.com/resources/",
+    hrefName: "Security Best Practices",
+  },
+  {
+    iconName: "github",
+    href: "https://github.com/gravitational/teleport",
+    hrefName: "Github Repo",
+  },
+];
 
 const DownloadSuccess = () => {
   return (
@@ -142,6 +173,40 @@ const StyledH3 = styled("h3")(
     lineHeight: ["md", "lg"],
     my: 0,
     pt: [5, 6],
+  })
+);
+
+interface LIProps {
+  href: string;
+  hrefName: string;
+  iconName?: IconName;
+}
+
+const LI = ({ href, hrefName, iconName }: LIProps) => {
+  return (
+    <StyledLi>
+      <StyledIcon name={iconName} mr={iconName ? 3 : -4} />
+      <Link color="darkest" href={href}>
+        {hrefName}
+      </Link>
+    </StyledLi>
+  );
+};
+
+const StyledLi = styled("li")(
+  css({
+    display: "flex",
+    lineHeight: "md",
+    pt: [2, 4],
+    "&:nth-child(1)": {
+      pt: [2, 3],
+    },
+  })
+);
+
+const StyledIcon = styled(Icon)(
+  css({
+    color: "dark-purple",
   })
 );
